Update to latest Rust and fix Clippy warnings

+++ benches/nizk.rs
@@ -18,7 +18,7 @@ fn nizk_prove_benchmark(c: &mut Criterion) {
     let mut group = c.benchmark_group("NIZK_prove_benchmark");
     group.plot_config(plot_config);
 
-    let num_vars = (2 as usize).pow(s as u32);
+    let num_vars = (2_usize).pow(s as u32);
     let num_cons = num_vars;
     let num_inputs = 10;
 
@@ -49,7 +49,7 @@ fn nizk_verify_benchmark(c: &mut Criterion) {
     let mut group = c.benchmark_group("NIZK_verify_benchmark");
     group.plot_config(plot_config);
 
-    let num_vars = (2 as usize).pow(s as u32);
+    let num_vars = (2_usize).pow(s as u32);
     let num_cons = num_vars;
     let num_inputs = 10;
     let (inst, vars, inputs) = Instance::produce_synthetic_r1cs(num_cons, num_vars, num_inputs);
